Refactor tab and contentscript functions
